# static/meta/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for LibRate, a social media cataloguing and reviewing service",
        "title": "LibRate",
        "contact": {
            "name": "MJH",
            "email": "TODO@flagship.instance"
        },
        "license": {
            "name": "GNU Affero General Public License v3",
            "url": "https://www.gnu.org/licenses/agpl-3.0.html"
        },
        "version": "dev"
    },
    "basePath": "/api",
    "paths": {
        "/artists/by-name": {
            "post": {
                "description": "Retrieve the artists with the given names",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media",
                    "artists",
                    "bulk operations"
                ],
                "summary": "Retrieve artists",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Artist names",
                        "name": "names",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ResponseHTTP"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.GroupedArtists"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    }
                }
            }
        },
        "/authenticate/login": {
            "post": {
                "description": "Create a session for the user",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth",
                    "accounts"
                ],
                "summary": "Login to the application",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Member name. Request must include either membername or email",
                        "name": "membername",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Email address",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "maximum": 2147483647,
                        "minimum": 1,
                        "type": "integer",
                        "default": 30,
                        "description": "Session time in minutes",
                        "name": "session_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Referrer-Policy header",
                        "name": "Referrer-Policy",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "X-CSRF-Token header",
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ResponseHTTP"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/auth.SessionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    }
                }
            }
        },
        "/genre/{kind}/{genre}": {
            "get": {
                "description": "Retrieve the genre with the given name and type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media",
                    "genres"
                ],
                "summary": "Retrieve genre",
                "parameters": [
                    {
                        "enum": [
                            "film",
                            "tv",
                            "music",
                            "book",
                            "game"
                        ],
                        "type": "string",
                        "description": "Genre kind",
                        "name": "kind",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Genre name (snake_lowercase)",
                        "name": "genre",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "en",
                            "de"
                        ],
                        "type": "string",
                        "description": "ISO-639-1 language code",
                        "name": "lang",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ResponseHTTP"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Genre"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    }
                }
            }
        },
        "/genres/{kind}": {
            "get": {
                "description": "Retrieve the list of genres of the specified type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media",
                    "genres",
                    "bulk operations"
                ],
                "summary": "Retrieve genres",
                "parameters": [
                    {
                        "enum": [
                            "film",
                            "tv",
                            "music",
                            "book",
                            "game"
                        ],
                        "type": "string",
                        "description": "Genre kind",
                        "name": "kind",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Return only genre names. Usually used for populating dropdowns",
                        "name": "names_only",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Return the genre names as links",
                        "name": "as_links",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Return all genres, not only the ones without a parent genre (e.g. Twee Pop and Jangle Pop instead of just Pop)",
                        "name": "all",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "name",
                                "id",
                                "kinds",
                                "parent",
                                "children"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Return only the specified columns",
                        "name": "columns",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "If names_only=false and as_links=false",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ResponseHTTP"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Genre"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    }
                }
            }
        },
        "/media/{id}": {
            "get": {
                "description": "Retrieve complete media information for the given media ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media",
                    "metadata"
                ],
                "summary": "Retrieve media information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Media UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ResponseHTTP"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    }
                }
            }
        },
        "/media/{media_id}/cast": {
            "get": {
                "description": "Get the full cast and crew involved with the creation of the media with given ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media",
                    "artists",
                    "bulk operations",
                    "films",
                    "television",
                    "anime"
                ],
                "summary": "Get the cast of the media with given ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The UUID of the media to get the cast of",
                        "name": "media_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.ResponseHTTP"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Cast"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    }
                }
            }
        },
        "/media/{media_id}/images": {
            "get": {
                "description": "Retrieve the image paths for the media with the given ID",
                "consumes": [
                    "json text/plain"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "media",
                    "metadata",
                    "images"
                ],
                "summary": "Retrieve image paths",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Media UUID",
                        "name": "media_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Media UUID",
                        "name": "media_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    }
                }
            }
        },
...
        "/members/follow/requests/{id}": {
            "put": {
                "description": "Accept a follow request",
                "consumes": [
                    "application/json"
                ],
                "summary": "Accept a follow request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The accepter's JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "The ID of the follow request",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    }
                }
            },
            "delete": {
                "description": "Reject a follow request",
                "consumes": [
                    "application/json"
                ],
                "summary": "Reject a follow request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The rejecter's JWT",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "The webfinger of the member who requested to follow",
                        "name": "follower",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "The ID of the follow request",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResponseHTTP"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.SessionResponse": {
            "type": "object",
            "properties": {
                "membername": {
                    "type": "string",
                    "example": "lain"
                },
                "token": {
                    "type": "string",
                    "example": "[A-Za-z0-9]{37}.[A-Za-z0-9]{147}.L-[A-Za-z0-9]{24}_[A-Za-z0-9]{25}-zNjCwGMr-[A-Za-z0-9]{27}"
                }
            }
        },
        "handlers.ResponseHTTP": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "models.Cast": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "integer"
                },
                "actors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Person"
                    }
                },
                "directors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Person"
                    }
                }
            }
        },
...
        "models.GroupedArtists": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Group"
                    }
                },
                "individual": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Person"
                    }
                }
            }
        },
...
    }
}